grpc-estudo-caso/proto/servidor-python/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf>=4.24.0
